nexus_ingest/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0   # The server to run FastAPI
pydantic==2.6.0             # Data validation
pika==1.3.2                 # Library to talk to RabbitMQ (CloudAMQP)
python-dotenv==1.0.1        # To read the .env file
websockets==12.0            # For real-time data